fix: move circular test types to requests dir to fix module resolution on Python 3.10
