--- generators/python/tests/utils/typeddict_models/types/resources/types/requests/circular/__init__.py
@@ -1,7 +1,0 @@
-from .edit_json import EditJsonParams
-from .edit_object_json import EditObjectJsonParams
-
-__all__ = [
-    "EditJsonParams",
-    "EditObjectJsonParams",
-]
